Uplift theia dependencies from v1.34.2 to 1.37.2
Perform a side (minor) 'yarn upgrade' to keep the yarn.lock fully
aligned herein. This resulting yarn.lock is then based on going from
clean to having run the yarn command locally a few times.

Base this uplifted-to Theia version on community [1]'s below. Add a
clarifying step in README so the yarn commands work locally. Fix the
rebuild:browser scripting so it works with this and versus electron's.

[1] https://theia-ide.org/releases/

Signed-off-by: Marco Miller <marco.miller@ericsson.com>