[docker example] Reduce @theia extensions to the minimum required for a Trace Viewer appliance
There is not necessarily a perfect solution here - all depends on how the resulting Docker
image/appliance is intended to be used. Already, the fact that no support for vscode extension
is provided, makes this app not suitable for use in software development. If we go to the logical
conclusion, we can also remove a lot of the current @theia framework extensions and make the app
and resulting image smaller.

One further potential improvement could be to make the trace viewer Theia extension not depend on
@theia/preferences - currently the app pulls the relatively big @theia/monaco just for that. Instead,
e.g. a static stub implementation could be used, just to set hardcoded values for the few preferences
used in this app/appliance.

Note: some extra @theia extensions will be pulled as dependencies of the current list

Signed-off-by: Marc Dumais <marc.dumais@ericsson.com>